Started implementation.

+++ Alchemy/local_2_10K_connected_plus.py
@@ -215,7 +215,7 @@ class NetGIN(torch.nn.Module):
 
         x = x_6_r
 
-        x_all = self.set2set_all(x_all, data.batch_all)
+        x_all = self.set2set_all(x_all, data.batch_all.to(torch.int))
         x = self.set2set(x, data.batch)
 
         x = torch.cat([x, x_all], dim=-1)
